Refactor ridiculous portalling and non-sensical layouts and drawer logic
